change rule name and adopt feedback
