+++ docs/sample.tslint.json
@@ -42,7 +42,6 @@
       true,
       { "order": "fields-first" }
     ],
-    "mergeable-namespace": true,
     "new-parens": true,
     "no-angle-bracket-type-assertion": true,
     "no-any": false,
@@ -72,6 +71,7 @@
       true,
       "check-function-in-method"
       ],
+    "no-mergeable-namespace": true,
     "no-null-keyword": true,
     "no-reference": true,
     "no-require-imports": true,
