+++ src/tsconfig.json
@@ -76,7 +76,6 @@
         "rules/maxLineLengthRule.ts",
         "rules/memberAccessRule.ts",
         "rules/memberOrderingRule.ts",
-        "rules/mergeableNamespaceRule.ts",
         "rules/newParensRule.ts",
         "rules/noAngleBracketTypeAssertionRule.ts",
         "rules/noAnyRule.ts",
@@ -96,6 +95,7 @@
         "rules/noInferrableTypesRule.ts",
         "rules/noInternalModuleRule.ts",
         "rules/noInvalidThisRule.ts",
+        "rules/noMergeableNamespaceRule.ts",
         "rules/noNamespaceRule.ts",
         "rules/noNullKeywordRule.ts",
         "rules/noReferenceRule.ts",
