+++ test/tsconfig.json
@@ -73,7 +73,6 @@
         "../src/rules/maxLineLengthRule.ts",
         "../src/rules/memberAccessRule.ts",
         "../src/rules/memberOrderingRule.ts",
-        "../src/rules/mergeableNamespaceRule.ts",
         "../src/rules/newParensRule.ts",
         "../src/rules/noAngleBracketTypeAssertionRule.ts",
         "../src/rules/noAnyRule.ts",
@@ -93,6 +92,7 @@
         "../src/rules/noInferrableTypesRule.ts",
         "../src/rules/noInternalModuleRule.ts",
         "../src/rules/noInvalidThisRule.ts",
+        "../src/rules/noMergeableNamespaceRule.ts",
         "../src/rules/noNamespaceRule.ts",
         "../src/rules/noNullKeywordRule.ts",
         "../src/rules/noReferenceRule.ts",
